requirements.txt
requests
rish
pyfiglet
PyYAML
